llama-server/requirements.txt
fastapi
uvicorn
transformers
torch
safetensors
pydantic
huggingface_hub
python-dotenv
Pillow
